Remove SqlQueryProviderModule and update tests
SqlQueryProviderModule and its corresponding test have been removed.
SqlFinderInterface has been created. SqlFinder class now implements SqlFinderInterface.
Minor changes to other test files and classes have been made, including adjustments to method parameter types and additions of new classes.
The changes are made to improve code structure and readability.
