+++ src/RowListInterfaceProvider.php
@@ -24,7 +24,7 @@ final class RowListInterfaceProvider implements ProviderInterface
     /** @var ExtendedPdoInterface */
     private $pdo;
 
-    /** @var SqlFinder */
+    /** @var SqlFinderInterface */
     private $finder;
 
     /** @var InjectorInterface  */
@@ -33,7 +33,7 @@ final class RowListInterfaceProvider implements ProviderInterface
     public function __construct(
         InjectionPointInterface $ip,
         ExtendedPdoInterface $pdo,
-        SqlFinder $finder,
+        SqlFinderInterface $finder,
         InjectorInterface $injector
     ) {
         $this->ip = $ip;
