+++ src/SqlQueryModule.php
@@ -4,31 +4,19 @@ declare(strict_types=1);
 
 namespace Ray\Query;
 
-use FilesystemIterator;
+use Koriym\ParamReader\ParamReader;
+use Koriym\ParamReader\ParamReaderInterface;
 use Ray\Di\AbstractModule;
-use RecursiveDirectoryIterator;
-use RecursiveIteratorIterator;
-use RecursiveRegexIterator;
-use RegexIterator;
-use SplFileInfo;
-
-use function file_get_contents;
-use function pathinfo;
+use Ray\Di\Scope;
 
 class SqlQueryModule extends AbstractModule
 {
-    /** @var string */
+    /** @var string  */
     private $sqlDir;
 
-    /** @var callable */
-    private $getSql;
-
-    public function __construct(string $sqlDir, ?AbstractModule $module = null, ?callable $getSql = null)
+    public function __construct(string $sqlDir, ?AbstractModule $module = null)
     {
         $this->sqlDir = $sqlDir;
-        $this->getSql = $getSql ?? static function (SplFileInfo $fileInfo): string {
-            return (string) file_get_contents($fileInfo->getPathname());
-        };
 
         parent::__construct($module);
     }
@@ -39,55 +27,13 @@ class SqlQueryModule extends AbstractModule
     protected function configure()
     {
         $this->bind(SqlDir::class)->toInstance(new SqlDir($this->sqlDir));
-        /** @var SplFileInfo $fileInfo */
-        foreach ($this->files($this->sqlDir) as $fileInfo) {
-            $name = pathinfo((string) $fileInfo->getRealPath())['filename'];
-            $sqlId = 'sql-' . $name;
-            $this->bind(QueryInterface::class)->annotatedWith($name)->toConstructor(
-                SqlQueryRowList::class,
-                "sql=$sqlId",
-            );
-            $this->bindCallableItem($name, $sqlId);
-            $this->bindCallableList($name, $sqlId);
-
-            $sql = (string) ($this->getSql)($fileInfo);
-            $this->bind()->annotatedWith($sqlId)->toInstance($sql);
-        }
-    }
-
-    protected function bindCallableItem(string $name, string $sqlId): void
-    {
-        $this->bind(RowInterface::class)->annotatedWith($name)->toConstructor(
-            SqlQueryRow::class,
-            "sql=$sqlId",
-        );
-    }
-
-    protected function bindCallableList(string $name, string $sqlId): void
-    {
-        $this->bind()->annotatedWith($name)->toConstructor(
-            SqlQueryRowList::class,
-            "sql=$sqlId",
-        );
-        $this->bind(RowListInterface::class)->annotatedWith($name)->toConstructor(
-            SqlQueryRowList::class,
-            "sql=$sqlId",
-        );
-    }
-
-    /** @psalm-suppress ArgumentTypeCoercion */
-    private function files(string $dir): RegexIterator
-    {
-        return new RegexIterator(
-            new RecursiveIteratorIterator(
-                new RecursiveDirectoryIterator(
-                    $dir,
-                    FilesystemIterator::CURRENT_AS_FILEINFO | FilesystemIterator::KEY_AS_PATHNAME | FilesystemIterator::SKIP_DOTS,
-                ),
-                RecursiveIteratorIterator::LEAVES_ONLY,
-            ),
-            '/^.+\.sql$/',
-            RecursiveRegexIterator::MATCH,
-        );
+        $this->bind(SqlFinderInterface::class)->to(SqlFinder::class)->in(Scope::SINGLETON);
+        $this->bind(ParamReaderInterface::class)->to(ParamReader::class)->in(Scope::SINGLETON);
+        $this->bind(RowInterface::class)->toProvider(RowInterfaceProvider::class);
+        $this->bind(RowListInterface::class)->toProvider(RowListInterfaceProvider::class);
+        $this->bind(InvokeInterface::class)->toProvider(RowListInterfaceProvider::class);
+        $this->bind(QueryInterface::class)->toProvider(RowListInterfaceProvider::class);
+        // AOP
+        $this->install(new SqlQueryInterceptModule());
     }
 }
